requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
pydantic==2.5.0
PyMuPDF==1.23.8
numpy==1.24.3
python-dotenv==1.0.0
requests==2.31.0
aiofiles==23.2.1
google-generativeai==0.3.2
sentence-transformers==2.2.2
torch==2.1.0
transformers==4.35.0

